pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "companion"
version = "0.1.0"
description = "Local AI voice companion — Whisper STT → Ollama LLM → Kokoro TTS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "kokoro>=0.9.0",
    "httpx>=0.27.0",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
    "av>=12.0.0",
    "pip>=25.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
]

[project.scripts]
companion = "companion.app:main"
companion-web = "companion.app:main_web"

[tool.hatch.build.targets.wheel]
packages = ["src/companion"]
